Final value getter
