+++ appserver/tests/payara-samples/classloader-data-api/src/main/webapp/WEB-INF/web.xml
@@ -7,7 +7,7 @@
         <servlet-class>org.glassfish.jersey.servlet.ServletContainer</servlet-class>
         <init-param>
             <param-name>jersey.config.server.provider.packages</param-name>
-            <param-value>fish.payara.samples.classloaderdataapi</param-value>
+            <param-value>fish.payara.samples.classloaderdata.api</param-value>
         </init-param>
         <load-on-startup>1</load-on-startup>
     </servlet>
